install overhang function

+++ contrib/brl/bseg/bsgm/bsgm_error_checking.h
@@ -35,7 +35,7 @@ void bsgm_remove_shadow_overhang(
   vil_image_view<float>& disp_img,
   const vil_image_view<T>& img_tar,
   const vil_image_view<T>& img_ref,
-  const vgl_vector_2d<double>& sun_dir,
+  const vgl_vector_2d<float>& sun_dir,
   int shadow_high,
   int shadow_low,
   float shadow_gradient_thresh,
